+++ SpreadsheetUtility/SpreadsheetUtility.Infrastructure/Abstractions/IExampleFileProvider.cs
@@ -1,0 +1,10 @@
+using SpreadsheetUtility.Infrastructure.Models;
+
+namespace SpreadsheetUtility.Infrastructure.Abstractions;
+
+public interface IExampleFileProvider
+{
+    Task<IEnumerable<ExampleFileInfo>> GetAvailableFilesAsync();
+    Task<FileDownloadDto> GetFileAsync(string fileName);
+    Task<ExampleFileInfo?> GetFileInfoAsync(string fileName);
+}