+++ SpreadsheetUtility/SpreadsheetUtility.Infrastructure/ApiClients/AuthApiClient.cs
@@ -4,23 +4,23 @@
 // </auto-generated>
 //----------------------
 
-#pragma warning disable 108 // Disable "CS0108 '{derivedDto}.ToJson()' hides inherited member '{dtoBase}.ToJson()'. Use the new keyword if hiding was intended."
-#pragma warning disable 114 // Disable "CS0114 '{derivedDto}.RaisePropertyChanged(String)' hides inherited member 'dtoBase.RaisePropertyChanged(String)'. To make the current member override that implementation, add the override keyword. Otherwise add the new keyword."
-#pragma warning disable 472 // Disable "CS0472 The result of the expression is always 'false' since a value of type 'Int32' is never equal to 'null' of type 'Int32?'
-#pragma warning disable 612 // Disable "CS0612 '...' is obsolete"
-#pragma warning disable 649 // Disable "CS0649 Field is never assigned to, and will always have its default value null"
-#pragma warning disable 1573 // Disable "CS1573 Parameter '...' has no matching param tag in the XML comment for ...
-#pragma warning disable 1591 // Disable "CS1591 Missing XML comment for publicly visible type or member ..."
-#pragma warning disable 8073 // Disable "CS8073 The result of the expression is always 'false' since a value of type 'T' is never equal to 'null' of type 'T?'"
-#pragma warning disable 3016 // Disable "CS3016 Arrays as attribute arguments is not CLS-compliant"
-#pragma warning disable 8600 // Disable "CS8600 Converting null literal or possible null value to non-nullable type"
-#pragma warning disable 8602 // Disable "CS8602 Dereference of a possibly null reference"
-#pragma warning disable 8603 // Disable "CS8603 Possible null reference return"
-#pragma warning disable 8604 // Disable "CS8604 Possible null reference argument for parameter"
-#pragma warning disable 8625 // Disable "CS8625 Cannot convert null literal to non-nullable reference type"
-#pragma warning disable 8765 // Disable "CS8765 Nullability of type of parameter doesn't match overridden member (possibly because of nullability attributes)."
+#pragma warning disable 108
+#pragma warning disable 114
+#pragma warning disable 472
+#pragma warning disable 612
+#pragma warning disable 649
+#pragma warning disable 1573
+#pragma warning disable 1591
+#pragma warning disable 8073
+#pragma warning disable 3016
+#pragma warning disable 8600
+#pragma warning disable 8602
+#pragma warning disable 8603
+#pragma warning disable 8604
+#pragma warning disable 8625
+#pragma warning disable 8765
 
-namespace SpreadsheetUtility.UI.Web.Services.Generated
+namespace SpreadsheetUtility.Infrastructure.ApiClients
 {
     using System = global::System;
 
@@ -35,9 +35,9 @@ namespace SpreadsheetUtility.UI.Web.Services.Generated
         private static System.Lazy<Newtonsoft.Json.JsonSerializerSettings> _settings = new System.Lazy<Newtonsoft.Json.JsonSerializerSettings>(CreateSerializerSettings, true);
         private Newtonsoft.Json.JsonSerializerSettings _instanceSettings;
 
-    #pragma warning disable CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
+    #pragma warning disable CS8618
         public SpreadsheetUtilitiesAuthApiClient(System.Net.Http.HttpClient httpClient)
-    #pragma warning restore CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
+    #pragma warning restore CS8618
         {
             BaseUrl = "https://localhost:7076/";
             _httpClient = httpClient;
@@ -98,7 +98,6 @@ namespace SpreadsheetUtility.UI.Web.Services.Generated
 
                     var urlBuilder_ = new System.Text.StringBuilder();
                     if (!string.IsNullOrEmpty(_baseUrl)) urlBuilder_.Append(_baseUrl);
-                    // Operation Path: "initiateSession"
                     urlBuilder_.Append("initiateSession");
                     urlBuilder_.Append('?');
                     urlBuilder_.Append(System.Uri.EscapeDataString("eMail")).Append('=').Append(System.Uri.EscapeDataString(ConvertToString(eMail, System.Globalization.CultureInfo.InvariantCulture))).Append('&');
@@ -186,7 +185,6 @@ namespace SpreadsheetUtility.UI.Web.Services.Generated
 
                     var urlBuilder_ = new System.Text.StringBuilder();
                     if (!string.IsNullOrEmpty(_baseUrl)) urlBuilder_.Append(_baseUrl);
-                    // Operation Path: "getSession"
                     urlBuilder_.Append("getSession");
                     urlBuilder_.Append('?');
                     urlBuilder_.Append(System.Uri.EscapeDataString("eMail")).Append('=').Append(System.Uri.EscapeDataString(ConvertToString(eMail, System.Globalization.CultureInfo.InvariantCulture))).Append('&');
@@ -278,7 +276,6 @@ namespace SpreadsheetUtility.UI.Web.Services.Generated
 
                     var urlBuilder_ = new System.Text.StringBuilder();
                     if (!string.IsNullOrEmpty(_baseUrl)) urlBuilder_.Append(_baseUrl);
-                    // Operation Path: "updateSession"
                     urlBuilder_.Append("updateSession");
                     urlBuilder_.Append('?');
                     urlBuilder_.Append(System.Uri.EscapeDataString("eMail")).Append('=').Append(System.Uri.EscapeDataString(ConvertToString(eMail, System.Globalization.CultureInfo.InvariantCulture))).Append('&');
@@ -465,10 +462,6 @@ namespace SpreadsheetUtility.UI.Web.Services.Generated
             return result == null ? "" : result;
         }
     }
-
-    
-
-
 
     [System.CodeDom.Compiler.GeneratedCode("NSwag", "14.7.1.0 (NJsonSchema v11.6.1.0 (Newtonsoft.Json v13.0.0.0))")]
     public partial class ApiException : System.Exception