+++ SpreadsheetUtility/SpreadsheetUtility.Infrastructure/Repositories/DeveloperRepository.cs
@@ -1,0 +1,12 @@
+using SpreadsheetUtility.Domain.Models;
+using SpreadsheetUtility.Domain.Repositories;
+
+namespace SpreadsheetUtility.Infrastructure.Repositories;
+
+public class DeveloperRepository : IDeveloperRepository
+{
+    public Task<List<Developer>> GetAllAsync()
+    {
+        return Task.FromResult(new List<Developer>());
+    }
+}