+++ SpreadsheetUtility/SpreadsheetUtility.Library/Providers/IDateTimeProvider.cs
@@ -1,9 +1,0 @@
-﻿namespace SpreadsheetUtility.Library.Providers
-{
-    public class DateTimeProvider : Application.Ports.IDateTimeProvider
-    {
-        public DateTime Now => DateTime.Now;
-        public DateTime Today => DateTime.Today;
-        public DateTime UtcNow => DateTime.UtcNow;
-    }
-}